fix non-linux builds
